# target/classes/openapi/api-out.yaml
openapi: 3.0.1
info:
  title: ValkyrAI CORE API
  description: The API for accessing ValkyrAI services
  termsOfService: https://valkyrlabs.com/tos/
  contact: {email: info@valkyrlabs.com}
  license: {name: AGPL 3.0, url: 'https://www.gnu.org/licenses/agpl-3.0.html'}
  version: 0.9.25
...
servers:
- {url: 'http://localhost:8080/v1', description: local server (uses live data)}
- {url: 'https://valkyrlabs.com/v1', description: production server}
...
paths:
...
  /ContentMediaLink:
    get:
      tags: [ContentMediaLink]
      summary: Retrieve a list of ContentMediaLinks
      description: Retrieves a list of ContentMediaLinks.
      x-spring-cacheable: true
      operationId: getContentMediaLinkList
      responses:
        '200':
          description: Successfully retrieved the list of ContentMediaLinks.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ContentMediaLink'}
        '404': {description: No ContentMediaLinks were found.}
        '500': {description: Internal server error.}
    post:
      tags: [ContentMediaLink]
      summary: Create a new ContentMediaLink
      description: Creates a new ContentMediaLink.
      operationId: postContentMediaLink
      requestBody:
        description: ContentMediaLink details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ContentMediaLink'}
      responses:
        '201':
          description: ContentMediaLink created successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ContentMediaLink'}
        '400': {description: Invalid input.}
        '404': {description: ContentMediaLink not found.}
        '500': {description: Internal server error.}
  /ContentMediaLink/{id}:
    get:
      tags: [ContentMediaLink]
      summary: Retrieve a single ContentMediaLink
      description: Retrieves a single ContentMediaLink for a specific uid.
      x-spring-cacheable: true
      operationId: getContentMediaLink
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the ContentMediaLink.
      responses:
        '200':
          description: Successfully retrieved the ContentMediaLink.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ContentMediaLink'}
        '404': {description: ContentMediaLink not found.}
        '500': {description: Internal server error.}
    put:
      tags: [ContentMediaLink]
      summary: Update an existing ContentMediaLink
      description: Updates an existing ContentMediaLink.
      operationId: updateContentMediaLink
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the ContentMediaLink.
      requestBody:
        description: Updated ContentMediaLink details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ContentMediaLink'}
      responses:
        '200':
          description: ContentMediaLink updated successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ContentMediaLink'}
        '400': {description: Invalid input.}
        '404': {description: ContentMediaLink not found.}
        '500': {description: Internal server error.}
    delete:
      tags: [ContentMediaLink]
      summary: Delete a ContentMediaLink.
      description: Deletes a specific ContentMediaLink.
      operationId: deleteContentMediaLink
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the ContentMediaLink.
      responses:
        '204': {description: ContentMediaLink deleted successfully.}
        '404': {description: ContentMediaLink not found.}
        '500': {description: Internal server error.}
...
components:
...
  schemas:
...
    ContentMediaLink:
      type: object
      description: link to images, videos, audio, and other content media
      required: [contentId]
      properties:
        applicationId: {type: string, format: uuid, description: used when linked
            to an Application}
        contentDataId: {type: string, format: uuid, description: used when linked
            to ContentData}
        contentUse:
          type: string
          description: specialized usage for the media object if any
          enum: [none, screenshot, selfie, portrait, promotional, informational, entertainment,
            product, support, training]
        contentType:
          type: string
          description: content type of the media object
          enum: [image, icon, video, audio, text, json, yaml, xml, markdown, spreadsheet,
            pdf]
        mediaUrl: {type: string, format: url, description: url to the media object}
        id:
          description: Unique identifier for object in the system
          type: string
          x-field-extra-annotation: "@Id  \n        @GeneratedValue(generator = \"\
            UUID\")\n"
          format: uuid
          example: db3b4973-11f0-4789-a296-c612f0c5e4a0
        ownerId: {description: UUID of owner of the object in the system, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.CREATED_BY,
            enabled = true)', format: uuid, example: 209291e8-4da1-459c-b369-03cd93b15707}
        createdDate: {description: Date of object creation, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.CREATED_DATE, enabled = true)', format: date-time,
          example: '2025-01-03T16:21:41.595-0800'}
        keyHash: {description: 'Data, including hash of the key(s) used to encrypt
            this record.', type: string}
        lastAccessedById: {description: Last user to access object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_BY,
            enabled = true)', format: uuid, example: e37f321d-d34f-4516-907b-f4626e955645}
        lastAccessedDate: {description: Timestamp of last access of object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.595-0800'}
        lastModifiedById: {description: Unique identifier for user who last modifed
            the object in the system, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.LAST_MODIFIED_BY, enabled = true)', format: uuid,
          example: 85b9ccff-ee11-4973-bbef-0082a6886e2e}
        lastModifiedDate: {description: Date of last object modification, type: string,
          x-field-extra-annotation: '    @AuditingField(fieldType = AuditingField.FieldType.LAST_MODIFIED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.595-0800'}
...
  securitySchemes:
    thorapi_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://{username}.valkyrlabs.com:{port}/oauth
          scopes: {'write:items': modify owned items, 'read:items': read owned items}